+++ apps/dashboard/netlify.toml
@@ -9,12 +9,6 @@
   ignore = "git diff --quiet $CACHED_COMMIT_REF $COMMIT_REF apps/dashboard"
 
 [[redirects]]
-  from = "/templates-proxy/*"
-  to = "https://templates-eta-self.vercel.app/:splat"
-  status = 200
-  force = true
-
-[[redirects]]
   from = "/legacy/*"
   to = "/legacy/index.html"
   status = 200
