+++ apps/dashboard/src/api/workflows.ts
@@ -105,24 +105,12 @@ export async function triggerWorkflow({
   payload: unknown;
   to: unknown;
 }) {
-  const basePayload: Record<string, unknown> =
-    payload !== null && typeof payload === 'object' && !Array.isArray(payload)
-      ? (payload as Record<string, unknown>)
-      : {};
-  const existingSource =
-    typeof (basePayload as { __source?: unknown }).__source === 'string'
-      ? (basePayload as { __source: string }).__source
-      : undefined;
-  const payloadWithSource: Record<string, unknown> = existingSource
-    ? basePayload
-    : { ...basePayload, __source: 'dashboard' };
-
   return post<{ data: { transactionId?: string } }>(`/events/trigger`, {
     environment,
     body: {
       name,
       to,
-      payload: payloadWithSource,
+      payload: { ...(payload ?? {}), __source: (payload as any)?.__source ?? 'dashboard' },
     },
   });
 }
